test: move BackgroundProcess to a utility file
